Move session into separate file
