+++ neqo-http3/src/features/extended_connect/tests/webtransport/sessions.rs
@@ -13,7 +13,7 @@ use crate::{
         tests::webtransport::{
             assert_wt, default_http3_client, default_http3_server, wt_default_parameters, WtTest,
         },
-        SessionCloseReason,
+        CloseReason,
     },
     frames::WebTransportFrame,
     Error, Header, Http3ClientEvent, Http3OrWebTransportStream, Http3Server, Http3ServerEvent,
@@ -33,11 +33,7 @@ fn wt_session_reject() {
     let accept_res = SessionAcceptAction::Reject(headers.clone());
     let (wt_session_id, _wt_session) = wt.negotiate_wt_session(&accept_res);
 
-    wt.check_session_closed_event_client(
-        wt_session_id,
-        &SessionCloseReason::Status(404),
-        Some(&headers),
-    );
+    wt.check_session_closed_event_client(wt_session_id, &CloseReason::Status(404), Some(&headers));
 }
 
 #[test]
@@ -46,10 +42,7 @@ fn wt_session_close_client() {
     let wt_session = wt.create_wt_session();
 
     wt.cancel_session_client(wt_session.stream_id());
-    wt.check_session_closed_event_server(
-        &wt_session,
-        &SessionCloseReason::Error(Error::HttpNone.code()),
-    );
+    wt.check_session_closed_event_server(&wt_session, &CloseReason::Error(Error::HttpNone.code()));
 }
 
 #[test]
@@ -60,7 +53,7 @@ fn wt_session_close_server() {
     wt.cancel_session_server(&wt_session);
     wt.check_session_closed_event_client(
         wt_session.stream_id(),
-        &SessionCloseReason::Error(Error::HttpNone.code()),
+        &CloseReason::Error(Error::HttpNone.code()),
         None,
     );
 }
@@ -74,7 +67,7 @@ fn wt_session_close_server_close_send() {
     wt.exchange_packets();
     wt.check_session_closed_event_client(
         wt_session.stream_id(),
-        &SessionCloseReason::Clean {
+        &CloseReason::Clean {
             error: 0,
             message: String::new(),
         },
@@ -93,7 +86,7 @@ fn wt_session_close_server_stop_sending() {
     wt.exchange_packets();
     wt.check_session_closed_event_client(
         wt_session.stream_id(),
-        &SessionCloseReason::Error(Error::HttpNone.code()),
+        &CloseReason::Error(Error::HttpNone.code()),
         None,
     );
 }
@@ -109,7 +102,7 @@ fn wt_session_close_server_reset() {
     wt.exchange_packets();
     wt.check_session_closed_event_client(
         wt_session.stream_id(),
-        &SessionCloseReason::Error(Error::HttpNone.code()),
+        &CloseReason::Error(Error::HttpNone.code()),
         None,
     );
 }
@@ -176,11 +169,7 @@ fn wt_session_response_with_redirect() {
 
     let (wt_session_id, _wt_session) = wt.negotiate_wt_session(&accept_res);
 
-    wt.check_session_closed_event_client(
-        wt_session_id,
-        &SessionCloseReason::Status(302),
-        Some(&headers),
-    );
+    wt.check_session_closed_event_client(wt_session_id, &CloseReason::Status(302), Some(&headers));
 }
 
 #[test]
@@ -222,7 +211,7 @@ fn wt_session_respone_200_with_fin() {
                 ..
             }) if (
                 stream_id == wt_session_id &&
-                reason == SessionCloseReason::Clean{ error: 0, message: String::new()} &&
+                reason == CloseReason::Clean{ error: 0, message: String::new()} &&
                 headers.is_none()
             )
         )
@@ -244,7 +233,7 @@ fn wt_session_close_frame_client() {
 
     wt.check_session_closed_event_server(
         &wt_session,
-        &SessionCloseReason::Clean {
+        &CloseReason::Clean {
             error: ERROR_NUM,
             message: ERROR_MESSAGE.to_string(),
         },
@@ -263,7 +252,7 @@ fn wt_session_close_frame_server() {
 
     wt.check_session_closed_event_client(
         wt_session.stream_id(),
-        &SessionCloseReason::Clean {
+        &CloseReason::Clean {
             error: ERROR_NUM,
             message: ERROR_MESSAGE.to_string(),
         },
@@ -313,7 +302,7 @@ fn wt_unknown_session_frame_client() {
         Some(Error::HttpRequestCancelled.code()),
         Some(&(
             wt_session.stream_id(),
-            SessionCloseReason::Clean {
+            CloseReason::Clean {
                 error: ERROR_NUM,
                 message: ERROR_MESSAGE.to_string(),
             },
@@ -342,12 +331,12 @@ fn wt_close_session_frame_broken_client() {
     // check that the webtransport session is closed.
     wt.check_session_closed_event_client(
         wt_session.stream_id(),
-        &SessionCloseReason::Error(Error::HttpGeneralProtocolStream.code()),
+        &CloseReason::Error(Error::HttpGeneralProtocolStream.code()),
         None,
     );
     wt.check_session_closed_event_server(
         &wt_session,
-        &SessionCloseReason::Error(Error::HttpGeneralProtocolStream.code()),
+        &CloseReason::Error(Error::HttpGeneralProtocolStream.code()),
     );
 
     // The Http3 session is still working.
@@ -427,7 +416,7 @@ fn wt_close_session_cannot_be_sent_at_once() {
         false,
         Some(&(
             wt_session.stream_id(),
-            SessionCloseReason::Clean {
+            CloseReason::Clean {
                 error: ERROR_NUM,
                 message: ERROR_MESSAGE.to_string(),
             },
